requirements.txt
antlr4-python3-runtime~=4.8
anyio~=4.6.0
appdirs~=1.4.4
appnope~=0.1.4
argon2-cffi~=23.1.0
argon2-cffi-bindings~=21.2.0
arrow~=1.3.0
asttokens~=2.4.1
async-lru~=2.0.4
attrs~=24.2.0
babel~=2.16.0
backcall~=0.2.0
beautifulsoup4~=4.12.3
bleach~=6.1.0
cattrs~=24.1.2
certifi~=2024.8.30
cffi~=1.17.1
charset-normalizer~=3.3.2
click~=8.1.7
colorama~=0.4.6
comm~=0.2.2
cryptography~=43.0.1
debugpy~=1.8.5
decorator~=5.1.1
defusedxml~=0.7.1
exceptiongroup~=1.2.2
executing~=2.1.0
fastjsonschema~=2.20.0
fhir-pyrate~=0.2.1
fhirpathpy~=0.2.2
flatten-json~=0.1.14
fqdn~=1.5.1
gitdb~=4.0.11
GitPython~=3.1.43
h11~=0.14.0
httpcore~=1.0.5
httpx~=0.27.2
idna~=3.10
importlib_metadata~=8.5.0
ipykernel~=6.29.5
ipython~=8.27.0
ipython-genutils~=0.2.0
ipywidgets~=8.1.5
isoduration~=20.11.0
jedi~=0.19.1
Jinja2~=3.1.4
json5~=0.9.25
jsonpointer~=3.0.0
jsonschema~=4.23.0
jsonschema-specifications~=2023.12.1
jupyter~=1.1.1
jupyter-cache~=1.0.0
jupyter-console~=6.6.3
jupyter-events~=0.10.0
jupyter-lsp~=2.2.5
jupyter-server-mathjax~=0.2.6
jupyter_client~=8.6.3
jupyter_core~=5.7.2
jupyter_server~=2.14.2
jupyter_server_terminals~=0.5.3
jupyterlab~=4.2.5
jupyterlab_pygments~=0.3.0
jupyterlab_server~=2.27.3
jupyterlab_widgets~=3.0.13
markdown-it-py~=3.0.0
MarkupSafe~=2.1.5
matplotlib-inline~=0.1.7
mdurl~=0.1.2
mistune~=3.0.2
nbclassic~=1.1.0
nbclient~=0.10.0
nbconvert~=7.16.4
nbdime~=4.0.2
nbformat~=5.10.4
nest-asyncio~=1.6.0
notebook~=7.2.2
notebook_shim~=0.2.4
numpy~=1.26.4
overrides~=7.7.0
packaging~=24.1
pandas~=1.5.3
pandocfilters~=1.5.1
parso~=0.8.4
pexpect~=4.9.0
pickleshare~=0.7.5
platformdirs~=4.3.6
prometheus_client~=0.21.0
prompt_toolkit~=3.0.47
psutil~=6.0.0
ptyprocess~=0.7.0
pure_eval~=0.2.3
pycparser~=2.22
Pygments~=2.18.0
PyJWT~=2.9.0
pyrsistent~=0.20.0
python-dateutil~=2.9.0.post0
python-json-logger~=2.0.7
pytz~=2024.2
PyYAML~=6.0.2
pyzmq~=26.2.0
qtconsole~=5.6.0
QtPy~=2.4.1
referencing~=0.35.1
requests~=2.32.3
requests-cache~=0.9.8
rfc3339-validator~=0.1.4
rfc3986-validator~=0.1.1
rich~=13.8.1
rpds-py~=0.20.0
Send2Trash~=1.8.3
setuptools~=75.1.0
six~=1.16.0
smmap~=5.0.1
sniffio~=1.3.1
soupsieve~=2.6
SQLAlchemy~=2.0.35
stack-data~=0.6.3
tabulate~=0.9.0
terminado~=0.18.1
tinycss2~=1.3.0
tornado~=6.4.1
tqdm~=4.66.5
traitlets~=5.14.3
types-python-dateutil~=2.9.0.20240906
typing_extensions~=4.12.2
uri-template~=1.3.0
url-normalize~=1.4.3
urllib3~=2.2.3
wcwidth~=0.2.13
webcolors~=24.8.0
webencodings~=0.5.1
websocket-client~=1.8.0
widgetsnbextension~=4.0.13
zipp~=3.20.2
